move content-panels to folder by each type
