+++ apps/fishing-map/features/content-panel/ContentPanel.tsx
@@ -6,15 +6,15 @@ import { Spinner } from '@globalfishingwatch/ui-components'
 
 import { IS_CHATBOT_ENABLED } from 'data/config'
 import { useSidePanel } from 'features/content-panel/contentPanel.hooks'
-import DataTerminologyContent from 'features/content-panel/DataTerminologyContent'
-import InfoContainer from 'features/content-panel/InfoContainer'
-import UserDatasetContent from 'features/content-panel/UserDatasetContent'
-import UserGuideContent from 'features/content-panel/UserGuideContent'
+import DataTerminologyContent from 'features/content-panel/data-terminology/DataTerminologyContent'
+import DatasetInfoContainer from 'features/content-panel/datasets-info/DatasetInfoContainer'
+import UserDatasetInfoContainer from 'features/content-panel/datasets-info/UserDatasetInfoContainer'
+import UserGuideContent from 'features/content-panel/user-guide/UserGuideContent'
 import { Route } from 'routes/_app'
 
 import styles from './ContentPanel.module.css'
 
-const ChatContent = lazy(() => import('features/content-panel/ChatContent'))
+const ChatContent = lazy(() => import('features/content-panel/chat/ChatContent'))
 
 const MIN_PANEL_WIDTH = 320
 const MAX_PANEL_WIDTH = 800
@@ -106,8 +106,8 @@ function ContentPanel({
         onMouseDown={handleMouseDown}
       />
       {sidePanelContent === 'userGuide' && <UserGuideContent />}
-      {sidePanelContent === 'datasets' && <InfoContainer />}
-      {sidePanelContent === 'userDataset' && <UserDatasetContent />}
+      {sidePanelContent === 'datasets' && <DatasetInfoContainer />}
+      {sidePanelContent === 'userDataset' && <UserDatasetInfoContainer />}
       {sidePanelContent === 'dataTerminology' && <DataTerminologyContent />}
       {sidePanelContent === 'chat' && IS_CHATBOT_ENABLED && (
         <Suspense fallback={<Spinner />}>
