+++ apps/fishing-map/features/content-panel/chat/ChatContent.module.css
