+++ apps/fishing-map/features/content-panel/chat/chat-agent.ts
