+++ apps/fishing-map/features/content-panel/chat/navigate-tool.ts
