+++ apps/fishing-map/features/help/HelpHub.tsx
@@ -9,7 +9,7 @@ import { IS_CHATBOT_ENABLED } from 'data/config'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import { useAppDispatch } from 'features/app/app.hooks'
 import { useSidePanel } from 'features/content-panel/contentPanel.hooks'
-import { selectIsGFWUser, selectIsGuestUser } from 'features/user/selectors/user.selectors'
+import { selectIsGFWUser } from 'features/user/selectors/user.selectors'
 import { useIsClientHydrated } from 'hooks/ssr.hooks'
 
 import hintsConfig from './hints.content'
@@ -105,7 +105,7 @@ function HelpHub() {
             {t((t) => t.common.userGuide)}
           </button>
         </li>
-        {IS_CHATBOT_ENABLED && !isGFWUser && (
+        {IS_CHATBOT_ENABLED && isGFWUser && (
           <li>
             <button
               type="button"
